modify mypage and exception
