+++ src/main/webapp/resources/css/meetupDetail.css
@@ -365,7 +365,7 @@ aside {
 .search-area {
   width: 100%;
   height: 40px;
-  border: 2px solid #000;
+  border: 1px solid #ddd;
   border-radius: 50px;
 }
 
@@ -380,7 +380,7 @@ aside {
 
 #sortOrder {
   padding: 5px;
-  border: 2px solid #000;
+  border: 1px solid #ddd;
   border-radius: 50px;
 }
